本番前最終プルリクエスト

+++ src/com/github/unchama/gui/ranking/huntingexp/WeekHuntingExpRankingMenuManager.java
@@ -14,7 +14,7 @@ import com.github.unchama.sql.moduler.RankingTableManager.TimeType;
 public class WeekHuntingExpRankingMenuManager extends RankingMenuManager {
 	@Override
 	protected String getLore(double value) {
-		return "総狩猟経験値:" + (int)value;
+		return "総狩猟経験値:" + (long)value;
 	}
 
 	@Override
