+++ src/com/github/unchama/gui/ranking/mineblock/TotalMineBlockRankingMenuManager.java
@@ -14,7 +14,7 @@ public final class TotalMineBlockRankingMenuManager extends TotalRankingMenuMana
 
 	@Override
 	protected String getLore(double value) {
-		return "総整地量:" + (int)value;
+		return "総整地量:" + (long)value;
 	}
