+++ src/com/github/unchama/gui/ranking/mineblock/YearMineBlockRankingMenuManager.java
@@ -15,7 +15,7 @@ public class YearMineBlockRankingMenuManager extends RankingMenuManager {
 
 	@Override
 	protected String getLore(double value) {
-		return "総整地量:" + (int)value;
+		return "総整地量:" + (long)value;
 	}
 
 	@Override
